requirements.txt
tqdm
gitpython
kiutils
appdata
gitpython
fastapi
dash
flask
pfluent
pyyaml